refactor: 🔨 update

+++ code_extensions/install.sh
@@ -1,2 +1,4 @@
 ./install.py
-ln -s $(pwd)/ws_extensions.py /usr/local/bin/ws_extensions
+if [[ ! -f "/usr/local/bin/ws_extensions" ]]; then
+  ln -s $(pwd)/ws_extensions.py /usr/local/bin/ws_extensions
+fi
